// gabinet-kosmetyczny-frontend-develop/src/app/feature/employess/model/new-worker.interface.ts
export interface INewWorker {
    email?: string;
    password?: string;
    role?: string;
    name?: string;
    surname?: string;
    street?: string;
    code?: string;
    town?: string;
    birthday?: Date;
    phone?: string;
    pesel?: string;
    dateOfEmployment?: Date;
    certificates?: string;
    pacemaker?: boolean;
    hermophilia?: boolean;
    psoriasis?: boolean;
    allergies?: string;
    infectiousDiseases?: boolean;
    discoloration?: boolean;
    bloodCirculationDisorders?: boolean;
    herpes?: boolean;
    fever?: boolean;
    pregnancy?: boolean;
    weakness?: boolean;
}